Started moving implementation details out of the header file.
The pendulum swings back.  I had everything in a single header to make
it easy to re-use the code (and because there had been a couple
templates that had to be implemented in the header).  Now that I'm
making this into a proper Arduino library, I want to make the public
header easier to understand and use without seeing so much of the
implementation detail.