Added selectSource and experimented a with blocking query.
selectSource requires some extra work in the state machine because some
of the documentation suggests that it's important to wait about 200 ms
after changing the source before issuing more commands.

I'm not sure I'll stick with the blocking queries (current just
queryVolume), but it's the simplest approach from the client's point of
view.