Reintroduced hooks for callbacks to client code.
Backed out of the blocking query for the current volume.

Queries responses are now returned to the client via overrides of the
methods in SerialAudio::Hooks.  Likewise, asynchronous notifications
use the same mechanism.

One compromise is that the query responses are currently routed through
a single method in Hooks.  Breaking them into response-specific methods
makes the v-table larger, and that actually eats a lot of RAM.  (I
would have thought the v-table would reside in program memory and that
the object would contain a pointer to it.

I'm still churning on how and where to define the enumerations.
Arduino's preprocessing nonsense seems to thwart some standard usage of
using directives and declarations.  Nonetheless, I think this is getting
closer to where I want it to be.