Reduced queue size back to 4, and simplified the example sketch.
I suspect it's probably not good to call back into the audio module
from a hook.  Though I didn't prove it, I think that's led to a
couple stack overflows.  So, for now, the example sketch sets flags
that it checks later when not in the callback.

Reducing the queue size lowered the memory usage again.  I started
piping in the ability to signal an enqueuing failure when the queue is
full, but I'm not sure how I want the API to do that yet.