Moved m_tocheck into the spare flags in m_state.
This saves a byte per audio instance.

That led to little code restructing in onEvent, which helps separate
the handling of device discovery from the handling for general query
responses.

Added a delay whenever a device is inserted to allow for the module to
scan the newly inserted device without forcing an in-flight operation
from timeout out.